// lib/interfaces/publications.interface.ts
// Generated by https://quicktype.io

export interface PublicationsResponse {
  results: Results;
}

export interface Results {
  count: number;
  totalPages: number;
  currentPage: number;
  results: Publication[];
}

export interface Publication {
  id: string;
  user_id: string;
  publication_type_id: number;
  city_id: number;
  title: string;
  description: string;
  content: string;
  reference_link: string;
  created_at: string;
  updated_at: string;
  votes_count: number;
  user: User;
  publication_type: PublicationType;
  images: Image[];
  tags: PublicationType[];
}

export interface Image {
  publication_id: string;
  image_url: string;
  order: number;
  created_at: string;
  updated_at: string;
}

export interface PublicationType {
  id: number;
  name: string;
  description: string;
  created_at: string;
  updated_at: string;
  image_url?: null;
}

export interface User {
  first_name: string;
  last_name: string;
  image_url: null | string;
}